refactor: Improve code quality and type safety across encoder modules
- Remove unnecessary type casts (i32) from bitstream put_bits calls in scalefactor encoding
- Derive Default trait for ShineConfig struct and remove manual implementation
- Refactor samplerate and bitrate index lookup functions to use iterator-based approach
- Add comprehensive safety documentation for shine_encode_buffer_interleaved function
- Fix comment formatting in mdct.rs and quantization.rs (add missing doc comment markers)
- Simplify array indexing in MDCT aliasing reduction (band * 18 + 0 to band * 18)
- Remove unnecessary variable declarations in shine_outer_loop function
- Simplify error mapping in mp3_encoder.rs using function reference instead of closure
- Improve code readability and reduce boilerplate while maintaining functionality

+++ crate/src/bitstream.rs
@@ -235,25 +235,25 @@ fn encode_main_data(config: &mut ShineGlobalConfig) -> EncodingResult<()> {
             if gr == 0 || scfsi[0] == 0 {
                 for sfb in 0..6 {
                     let sf_val = config.scalefactor.l[gr][ch][sfb];
-                    config.bs.put_bits(sf_val as u32, slen1 as i32)?;
+                    config.bs.put_bits(sf_val as u32, slen1)?;
                 }
             }
             if gr == 0 || scfsi[1] == 0 {
                 for sfb in 6..11 {
                     let sf_val = config.scalefactor.l[gr][ch][sfb];
-                    config.bs.put_bits(sf_val as u32, slen1 as i32)?;
+                    config.bs.put_bits(sf_val as u32, slen1)?;
                 }
             }
             if gr == 0 || scfsi[2] == 0 {
                 for sfb in 11..16 {
                     let sf_val = config.scalefactor.l[gr][ch][sfb];
-                    config.bs.put_bits(sf_val as u32, slen2 as i32)?;
+                    config.bs.put_bits(sf_val as u32, slen2)?;
                 }
             }
             if gr == 0 || scfsi[3] == 0 {
                 for sfb in 16..21 {
                     let sf_val = config.scalefactor.l[gr][ch][sfb];
-                    config.bs.put_bits(sf_val as u32, slen2 as i32)?;
+                    config.bs.put_bits(sf_val as u32, slen2)?;
                 }
             }
